requirements.txt
asgiref==3.8.1
Django==5.1.1
django-bootstrap5==24.3
sqlparse==0.5.1
tzdata==2024.2
gunicorn
python-decouple
psycopg[binary]
dj-database-url
requests
whitenoise
django_allauth[socialaccount]
django-allauth-ui
django-widget-tweaks
slippers
stripe
djangorestframework
django-environ
PyJWT
cryptography